Refactor(#144): delayArr & ROWS_PER_PAGE를 외부로 이동
