+++ app/artwork/lib/const.ts
@@ -1,0 +1,1 @@
+export const ROWS_PER_PAGE = 30; // 한 페이지당 불러올 아이템 개수
